fix: update stat property types

+++ src/Objects/Stat.php
@@ -8,29 +8,29 @@ use stdClass;
 
 class Stat extends TypesenseObject
 {
-    public int $delete_latency_ms;
+    public float $delete_latency_ms;
 
-    public int $delete_requests_per_second;
+    public float $delete_requests_per_second;
 
-    public int $import_latency_ms;
+    public float $import_latency_ms;
 
-    public int $import_requests_per_second;
+    public float $import_requests_per_second;
 
     public stdClass $latency_ms;
 
-    public int $overloaded_requests_per_second;
+    public float $overloaded_requests_per_second;
 
-    public int $pending_write_batches;
+    public float $pending_write_batches;
 
     public stdClass $requests_per_second;
 
-    public int $search_latency_ms;
+    public float $search_latency_ms;
 
-    public int $search_requests_per_second;
+    public float $search_requests_per_second;
 
     public float $total_requests_per_second;
 
-    public int $write_latency_ms;
+    public float $write_latency_ms;
 
-    public int $write_requests_per_second;
+    public float $write_requests_per_second;
 }
